+++ geotrek/api/v2/views/signage.py
@@ -1,0 +1,46 @@
+from django.conf import settings
+from django.db.models import F
+from geotrek.api.v2 import serializers as api_serializers, \
+    viewsets as api_viewsets, filters as api_filters
+from geotrek.signage import models as signage_models
+from geotrek.api.v2.functions import Transform
+
+
+class SignageViewSet(api_viewsets.GeotrekGeometricViewset):
+    filter_backends = api_viewsets.GeotrekGeometricViewset.filter_backends + (api_filters.NearbyContentFilter, api_filters.UpdateOrCreateDateFilter)
+    serializer_class = api_serializers.SignageSerializer
+    queryset = signage_models.Signage.objects.existing() \
+        .select_related('topo_object', 'type', ) \
+        .annotate(geom3d_transformed=Transform(F('geom_3d'), settings.API_SRID)) \
+        .prefetch_related('topo_object__aggregations', 'attachments') \
+        .order_by('pk')
+
+
+class SignageTypeViewSet(api_viewsets.GeotrekViewSet):
+    filter_backends = api_viewsets.GeotrekViewSet.filter_backends + (api_filters.SignageRelatedPortalFilter, )
+    serializer_class = api_serializers.SignageTypeSerializer
+    queryset = signage_models.SignageType.objects.all().order_by('pk')
+
+
+class DirectionViewSet(api_viewsets.GeotrekViewSet):
+    filter_backends = api_viewsets.GeotrekViewSet.filter_backends
+    serializer_class = api_serializers.DirectionSerializer
+    queryset = signage_models.Direction.objects.all().order_by('pk')
+
+
+class SealingViewSet(api_viewsets.GeotrekViewSet):
+    filter_backends = api_viewsets.GeotrekViewSet.filter_backends
+    serializer_class = api_serializers.SealingSerializer
+    queryset = signage_models.Sealing.objects.all().order_by('pk')
+
+
+class ColorViewSet(api_viewsets.GeotrekViewSet):
+    filter_backends = api_viewsets.GeotrekViewSet.filter_backends
+    serializer_class = api_serializers.ColorSerializer
+    queryset = signage_models.Color.objects.all().order_by('pk')
+
+
+class BladeTypeViewSet(api_viewsets.GeotrekViewSet):
+    filter_backends = api_viewsets.GeotrekViewSet.filter_backends
+    serializer_class = api_serializers.BladeTypeSerializer
+    queryset = signage_models.BladeType.objects.all().order_by('pk')